spotify_downloader/requirement.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
Django==6.0.3
downloader==0.98
Flask==3.1.3
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
six==1.17.0
sniffio==1.3.1
spotify-api==0.0.8
spotipy==2.26.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.8
youtube-search==1.1.1
